Add initial implementation of Tile
This comes with a number of related classes.

The Direction class allows to represent a given direction, in a hexagonal grid.
A Tile class represents an hexagonal tile in the map.

The DirectionnedTile has a single purpose: grouping toghether a tile and a
direction. This allows us to have a nice API to create tiles with the
Tile:neighboorOf static function.