Add pattern matcher, pattern comparator
The pattern class is responsible for finding specific patterns in the game map.
A Pattern is represented as a constraint for a specific tile, as well as its six
neighbors.

The equals and hashCode methods have been tweaked so that two pattern are equals
if there is an angle at which their constraints are equals. HashCode follows the
same principle.

The Pattern class can also search in a Map, and find where this matches. This
method will return a set of MatchResult. A MatchResult represents a Pattern that
has successfully matched at a given coordinate. This also allows us to remove a
huge amount of possibly-duplicated matches.