Remove sideLen attribute of Coordinate
Previous implementation of Coordinate featured a sideLen field, which represented
how much tiles the map has on each side. The downside of this was that the
constructor of Coordinate had three arguments, which is non-obvious, given that it
represents a (x, y) coordinate.

This commit abstracts away this notion of side length. It is now a parameter of the
toOffset method, and is exclusively stored in the Map instead.