Split Irrigation/IrrigationCoordinate
For tiles, we have a clear dinstinction between Tile and Coordinate: a Tile
contains a Coordinate, but a Coordinate does not represent a Tile in itself.
This commit introduces a similar scheme for Irrigation and IrrigationCoordinate.
An IrrigationCoordinate represents somewhere an irrigation can be placed on the
map, but not the irrigation itself.