Rename and split Coordinate::fromNeighbors
There are now two functions.

The first one, Coordinate::fromNeighbors, compute the converging coordinate, check
that it is correct, and return it.
The second one, Coordinate::validFromNeighbors, calls Coordinate::fromNeighbors,
and checks that the resulting coordinate has the correct amount of neighbors.

This will allow us to delay the neighbor check in Map::addNeighborOf.