Rework equality system to better handle the null pointer case
Previous implementation of equals generated an IllegalComparaisonException when a non-correct
type was being compared with another. The problem is that this exception was generated by calling
.getClass() on the right-hand-side object, which can be null.

This commit handles the null pointer case by requiring the object to be passed to
IllegalEqualityExceptionGenerator without any transformation. The null case is then handled in
the corresponding class.